+++ packages/angular/src/index.ts
@@ -2,4 +2,5 @@
  * @3plate/graph-angular - Angular components for @3plate/graph graph visualization
  */
 
-// Placeholder for Angular integration
+export { GraphComponent } from './graph.component'
+export { PlaygroundComponent } from './playground.component'