+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @3plate/graph-core
 
+## 0.1.9
+
+### Patch Changes
+
+- Add react, vue, angular wrappers
+
 ## 0.1.8
 
 ### Patch Changes